feat: Create ISubscribe interface
